requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
replicate
supabase
PyJWT
